+++ python/pyemd_rs/__init__.py
@@ -2,7 +2,12 @@
 
 from __future__ import annotations
 
-from ._pyemd_rs import ceemdan, emd
+from ._pyemd_rs import ceemdan, default_ceemdan_opts, default_emd_opts, emd
 
 __version__ = "0.1.3"
-__all__ = ["ceemdan", "emd"]
+__all__ = [
+    "ceemdan",
+    "default_ceemdan_opts",
+    "default_emd_opts",
+    "emd",
+]